requirements.txt
python-telegram-bot[job-queue,rate-limiter]>=20.0
requests>=2.31.0
cryptography>=41.0.0
